Fix possible UnicodeDecodeError in warning when msgid already exists.
This happens when comparing 'Read more&hellip;' with 'Read more…'.
That should not give a warning, because they are really the same, but that would need a fix somewhere else,
in the code that gets the value of an i18n:attribute in a different way than an i18n:translate.

Added test for the proper printing of the warning.