Fix tests for Python 3.13.6+ compatibility. (#22)
Python 3.13.6 changed the way `<title>` tags are parsed: If there is no end tag,
the rest of document is handled as title content.
Source: https://github.com/python/cpython/pull/135310

This change has also been backported to Python 3.12 but is not yet released.

Example of a breaking test: https://github.com/zopefoundation/groktoolkit/actions/runs/16882111176/job/47820322945?pr=102
(Sorry I was not able to reproduce it here in the repository but I also saw it locally on 3.13.6)  – The difference is that groktoolkit already uses uv (which already has 3.13.6) and zope.tal still relies on the Python versions provided by GitHub Actions (which does not yet have it.)