tests: resolve symlinks for path normalization
tests/input/test_sa4.html includes relative path to
```
<p metal:use-macro="tests/input/test_sa3.html/macro1">
```

which is then normalized with os.path.abspath
(i.e. normpath(join(os.getcwd(), path)), this is the resolved path)
and removing a prefix (see zope.tal.driver.compilefile for details).

It's expected to be normalized to tests/input/test_sa3.html but actual value
differs for venv-based envs, for example:
/usr/src/RPM/BUILD/python3-module-zope.tal-5.1.1/.run_venv/lib/python3/site-packages/zope/tal/tests/input/test_sa3.html

This happens because in stdlib's `venv` the `lib64` directory is a symlink to
`lib`, while in `virtualenv` (used by upstream) `lib64` is a regular
directory.

Closes #25
Signed-off-by: Stanislav Levin <slev@altlinux.org>